move test_vocab_download_glove_vectors (test.test_vocab.TestVocab)
